refactored into ResponseHandler2019 and ResponseHandler2023 files
