+++ src/aind_metadata_service/sharepoint/response_handler_2019.py
@@ -1,0 +1,699 @@
+"""Module to handle 2019 sharepoint response and map to procedures"""
+
+from enum import Enum
+from typing import Optional
+
+from aind_data_schema.procedures import (
+    Anaesthetic,
+    Craniotomy,
+    CraniotomyType,
+    FiberImplant,
+    Headframe,
+    IontophoresisInjection,
+    NanojectInjection,
+    OphysProbe,
+    ProbeName,
+)
+from dateutil import parser
+from office365.runtime.client_object import ClientObject
+from office365.sharepoint.listitems.collection import ListItemCollection
+
+from aind_metadata_service.sharepoint.utils import (
+    convert_str_to_bool,
+    convert_str_to_time,
+    map_choice,
+    map_hemisphere,
+    parse_str_into_float,
+)
+
+
+class NeurosurgeryAndBehaviorList2019:
+    """Class to contain helpful info to parse the 2019 SharePoint List"""
+
+    class StringParserHelper(Enum):
+        """Enum class for SharePoint's response strings"""
+
+        # Not really why, but some response fields return 'Select...'
+        SELECT_STR = "Select..."
+        PROCEDURE_TYPE_SPLITTER = "+"
+
+    class ProcedureType(Enum):
+        """Enum class for SharePoint's Procedure Type"""
+
+        HEAD_PLANT = "HP"
+        STEREOTAXIC_INJECTION = "Stereotaxic Injection (Coordinate)"
+        INJECTION = "Injection"
+        INJ = "INJ"
+        OPTIC_FIBER_IMPLANT = "Optic Fiber Implant"
+        HP_TRANSCRANIAL = "HP Transcranial (for ISI)"
+        WHOLE_HEMISPHERE_CRANIOTOMY_NP = "WHC NP"
+        C_CAM = "C CAM"
+        C_MULTISCOPE = "C Multiscope"
+        C = "C"
+
+    class InjectionType(Enum):
+        """Enum class for Injection Types"""
+
+        IONTO = "Iontophoresis"
+        NANOJECT = "Nanoject (Pressure)"
+
+    class CraniotomyType(Enum):
+        """Enum class for Craniotomy Types"""
+
+        VISUAL_CORTEX = "Visual Cortex 5mm"
+        FRONTAL_WINDOW = "Frontal Window 3mm"
+        WHC_NP = "WHC NP"
+        WHC_2P = "WHC 2P"
+
+    class HeadPostType(Enum):
+        """Enum class for HeadPost Types"""
+
+        CAM = "CAM-style headframe (0160-100-10 Rev A)"
+        NEUROPIXEL = "Neuropixel-style headframe (0160-100-10/0160-200-36)"
+        MESO_NGC = (
+            "Mesoscope-style well with NGC-style headframe"
+            " (0160-200-20/0160-100-10)"
+        )
+        WHC_NP = "WHC #42 with Neuropixel well and well cap"
+        NGC = "NGC-style headframe, no well (0160-100-10)"
+        AI_HEADBAR = "AI Straight Headbar"
+
+    class ListField(Enum):
+        """Enum class for fields in List Item object response"""
+
+        FILE_SYSTEM_OBJECT_TYPE = "FileSystemObjectType"
+        ID = "Id"
+        SERVER_REDIRECTED_EMBED_URI = "ServerRedirectedEmbedUri"
+        SERVER_REDIRECTED_EMBED_URL = "ServerRedirectedEmbedUrl"
+        CONTENT_TYPE_ID = "ContentTypeId"
+        TITLE = "Title"
+        PI_ID = "PIId"
+        PI_STRING_ID = "PIStringId"
+        LAB_TRACKS_REQUESTOR = "LabTracks_x0020_Requestor"
+        PROJECT_ID_TE = "Project_x0020_ID_x0020__x0028_te"
+        IACUC_PROTOCOL = "IACUC_x0020_Protocol_x0020__x002"
+        PROCEDURE = "Procedure"
+        LIGHT_CYCLE = "Light_x0020_Cycle"
+        LI_MS_REQUIRED = "LIMs_x0020_Required"
+        LAB_TRACKS_GROUP = "LabTracks_x0020_Group"
+        LAB_TRACKS_ID = "LabTracks_x0020_ID"
+        DATE_OF_SURGERY = "Date_x0020_of_x0020_Surgery"
+        LIMS_PROJECT_CODE = "Lims_x0020_Project_x0020_Code"
+        VIRUS_M_L = "Virus_x0020_M_x002f_L"
+        VIRUS_A_P = "Virus_x0020_A_x002f_P"
+        VIRUS_D_V = "Virus_x0020_D_x002f_V"
+        VIRUS_HEMISPHERE = "Virus_x0020_Hemisphere"
+        HP_M_L = "HP_x0020_M_x002f_L"
+        HP_A_P = "HP_x0020_A_x002f_P"
+        HP_DIAMETER = "HP_x0020_Diameter"
+        ISO_ON = "Iso_x0020_On"
+        CAGE = "Cage"
+        SEX = "Sex"
+        DATE_OF_BIRTH = "Date_x0020_of_x0020_Birth"
+        WEIGHT_BEFORE_SURGER = "Weight_x0020_before_x0020_Surger"
+        WEIGHT_AFTER_SURGERY = "Weight_x0020_after_x0020_Surgery"
+        PEDIGREE_NAME = "PedigreeName"
+        BREG2_LAMB = "Breg2Lamb"
+        HEADPOST_TYPE = "HeadpostType"
+        DATE_RANGE_START = "DateRangeStart"
+        DATE_RANGE_END = "DateRangeEnd"
+        HP_LOC = "HpLoc"
+        HP_PERF = "HpPerf"
+        HP_DUROTOMY = "HPDurotomy"
+        HP_PREV_INJECT = "HpPrevInject"
+        ML2ND_INJ = "ML2ndInj"
+        AP2ND_INJ = "AP2ndInj"
+        DV2ND_INJ = "DV2ndInj"
+        HEMISPHERE2ND_INJ = "Hemisphere2ndInj"
+        HP_WORK_STATION = "HpWorkStation"
+        SURGERY_STATUS = "SurgeryStatus"
+        COM_DUROTOMY = "ComDurotomy"
+        COM_SWELLING = "ComSwelling"
+        COM_SINUSBLEED = "ComSinusbleed"
+        COM_DURING1ST_INJ = "ComDuring1stInj"
+        COM_DURING2ND_INJ = "ComDuring2ndInj"
+        COM_DAMAGE = "ComDamage"
+        COM_WINDOW = "ComWindow"
+        COM_COPLANAR = "ComCoplanar"
+        COM_AFTER1ST_INJ = "ComAfter1stInj"
+        COM_AFTER2ND_INJ = "ComAfter2ndInj"
+        WORK_STATION1ST_INJECTION = "WorkStation1stInjection"
+        WORK_STATION2ND_INJECTION = "WorkStation2ndInjection"
+        DATE1ST_INJECTION = "Date1stInjection"
+        DATE2ND_INJECTION = "Date2ndInjection"
+        INJ1_STORAGE_LOCATION = "Inj1StorageLocation"
+        INJ2_STORAGE_LOCATION = "Inj2StorageLocation"
+        INJ1_TYPE = "Inj1Type"
+        INJ2_TYPE = "Inj2Type"
+        INJ1_VOL = "Inj1Vol"
+        INJ2_VOL = "Inj2Vol"
+        INJ1_LENGHTOF_TIME = "Inj1LenghtofTime"
+        INJ2_LENGHTOF_TIME = "Inj2LenghtofTime"
+        INJ1_CURRENT = "Inj1Current"
+        INJ2_CURRENT = "Inj2Current"
+        INJ1_ALTERNATING_TIME = "Inj1AlternatingTime"
+        INJ2_ALTERNATING_TIME = "Inj2AlternatingTime"
+        FIRST_INJECTION_WEIGHT_BEFOR = "FirstInjectionWeightBefor"
+        FIRST_INJECTION_WEIGHT_AFTER = "FirstInjectionWeightAfter"
+        FIRST_INJECTION_ISO_DURATION = "FirstInjectionIsoDuration"
+        SECOND_INJECTION_WEIGHT_BEFORE = "SecondInjectionWeightBefore"
+        SECOND_INJECTION_WEIGHT_AFTER = "SecondInjectionWeightAfter"
+        SECOND_INJECTION_ISO_DURATION = "SecondInjectionIsoDuration"
+        INJ1_ROUND = "Inj1Round"
+        INJ2_ROUND = "Inj2Round"
+        HP_ISO_LEVEL = "HPIsoLevel"
+        ROUND1_INJ_ISOLEVEL = "Round1InjIsolevel"
+        ROUND2_INJ_ISOLEVEL = "Round2InjIsolevel"
+        TEST1_ID = "Test1Id"
+        TEST1_STRING_ID = "Test1StringId"
+        TEST_2ND_ROUND_ID = "TEST_x0020_2nd_x0020_Round_x0020Id"
+        TEST_2ND_ROUND_STRING_ID = "TEST_x0020_2nd_x0020_Round_x0020StringId"
+        TEST_1ST_ROUND_ID = "TEST_x0020_1st_x0020_Round_x0020Id"
+        TEST_1ST_ROUND_STRING_ID = "TEST_x0020_1st_x0020_Round_x0020StringId"
+        ODATA_HP_REQUESTOR = "OData__x0031_HP_x0020_Requestor_x0020_"
+        ISSUE = "Issue"
+        TOUCH_UP_STATUS = "Touch_x0020_Up_x0020_Status"
+        TOUCH_UP_SURGEON_ID = "Touch_x0020_Up_x0020_SurgeonId"
+        TOUCH_UP_SURGEON_STRING_ID = "Touch_x0020_Up_x0020_SurgeonStringId"
+        TOUCH_UP_COMP = "Touch_x0020_Up_x0020__x0020_Comp"
+        EXUDATE_SEVERITY = "Exudate_x0020_Severity"
+        SCABBING = "Scabbing"
+        EYE_ISSUE = "Eye_x0020_Issue"
+        EYE_AFFECTED = "Eye_x0020_Affected"
+        TOUCH_UP_WEIGHT = "Touch_x0020_Up_x0020_Weight_x002"
+        LIMS_LINK = "LIMS_x0020_link"
+        HP_INJ = "HP_x0020__x0026__x0020_Inj"
+        FIELD30 = "field30"
+        FIELD50 = "field50"
+        LIM_STASKFLOW1 = "LIMStaskflow1"
+        COMPLIANCE_ASSET_ID = "ComplianceAssetId"
+        CREATED = "Created"
+        AUTHOR_ID = "AuthorId"
+        EDITOR_ID = "EditorId"
+        MODIFIED = "Modified"
+        HP_REQUESTOR_COMMENTS_PLAINTEXT = "HPRequestorCommentsPlaintext"
+        NANOJECT_NUMBER_INJ2 = "NanojectNumberInj2"
+        NANOJECT_NUMBER_INJ10 = "NanojectNumberInj10"
+        IONTO_NUMBER_INJ1 = "IontoNumberInj1"
+        IONTO_NUMBER_INJ2 = "IontoNumberInj2"
+        IONTO_NUMBER_HPINJ = "IontoNumberHPINJ"
+        INJ1VOLPERDEPTH = "inj1volperdepth"
+        INJ2VOLPERDEPTH = "inj2volperdepth"
+        INJ1ANGLE0 = "Inj1angle0"
+        INJ2ANGLE0 = "Inj2angle0"
+        CONTUSION = "Contusion"
+        HP_SURGEON_COMMENTS = "HPSurgeonComments"
+        ST_ROUND_INJECTION_COMMENTS = "stRoundInjectionComments"
+        ND_ROUNG_INJECTION_COMMENTS = "ndRoungInjectionComments"
+        FIRST_ROUND_IONTO_ISSUE = "FirstRoundIontoIssue"
+        HP_RECOVERY = "HPRecovery"
+        FIRST_INJ_RECOVERY = "FirstInjRecovery"
+        SECOND_INJ_RECOVER = "SecondInjRecover"
+        SECOND_ROUND_IONTO_ISSUE = "SecondRoundIontoIssue"
+        LONG_SURGEON_COMMENTS = "LongSurgeonComments"
+        LONG1ST_ROUND_INJ_CMTS = "Long1stRoundInjCmts"
+        LONG2ND_RND_INJ_CMTS = "Long2ndRndInjCmts"
+        LONG_REQUESTOR_COMMENTS = "LongRequestorComments"
+        INJ1_VIRUS_STRAIN_RT = "Inj1VirusStrain_rt"
+        INJ2_VIRUS_STRAIN_RT = "Inj2VirusStrain_rt"
+        RET_SETTING0 = "retSetting0"
+        RET_SETTING1 = "retSetting1"
+        START_OF_WEEK = "Start_x0020_Of_x0020_Week"
+        END_OF_WEEK = "End_x0020_of_x0020_Week"
+        AGE_AT_INJECTION = "Age_x0020_at_x0020_Injection"
+        CRANIOTOMY_TYPE = "CraniotomyType"
+        IMPLANT_ID_COVERSLIP_TYPE = "ImplantIDCoverslipType"
+        INJ1_ANGLE_V2 = "Inj1Angle_v2"
+        INJ2_ANGLE_V2 = "Inj2Angle_v2"
+        FIBER_IMPLANT1 = "FiberImplant1"
+        FIBER_IMPLANT1_DV = "FiberImplant1DV"
+        FIBER_IMPLANT2 = "FiberImplant2"
+        FIBER_IMPLANT2_DV = "FiberImplant2DV"
+        ID2 = "ID"  # For some reason ID and Id are present in response
+        ODATA_UI_VERSION_STRING = "OData__UIVersionString"
+        ATTACHMENTS = "Attachments"
+        GUID = "GUID"
+
+
+class ResponseHandler2019:
+    """Module to map sharepoint 2019 response"""
+    def handle_response_2019(self, list_items: ListItemCollection,
+                              head_frames, injections, fiber_implants, craniotomies):
+        """Handles sharepoint response when 2019 version"""
+        list_fields = NeurosurgeryAndBehaviorList2019.ListField
+        str_helpers = NeurosurgeryAndBehaviorList2019.StringParserHelper
+        nsb_proc_types = NeurosurgeryAndBehaviorList2019.ProcedureType
+        for list_item in list_items:
+            if list_item.get_property(list_fields.PROCEDURE.value):
+                procedure_types = list_item.get_property(
+                    list_fields.PROCEDURE.value
+                ).split(str_helpers.PROCEDURE_TYPE_SPLITTER.value)
+            else:
+                procedure_types = []
+            for procedure_type in procedure_types:
+                if procedure_type == nsb_proc_types.HEAD_PLANT.value:
+                    head_frames.append(
+                        self._map_list_item_to_head_frame(list_item)
+                    )
+                if procedure_type in {
+                    nsb_proc_types.STEREOTAXIC_INJECTION.value,
+                    nsb_proc_types.INJECTION.value,
+                    nsb_proc_types.INJ.value,
+                }:
+                    injections.append(
+                        self._map_list_item_to_injection(list_item)
+                    )
+                if (
+                        procedure_type
+                        == nsb_proc_types.OPTIC_FIBER_IMPLANT.value
+                ):
+                    fiber_implants.append(
+                        self._map_list_item_to_fiber_implant(list_item)
+                    )
+                if procedure_type in {
+                    nsb_proc_types.WHOLE_HEMISPHERE_CRANIOTOMY_NP.value,
+                    nsb_proc_types.C_MULTISCOPE.value,
+                    nsb_proc_types.C_CAM.value,
+                    nsb_proc_types.C.value,
+                }:
+                    craniotomies.append(
+                        self._map_list_item_to_craniotomy(list_item)
+                    )
+        return head_frames, injections, fiber_implants, craniotomies
+
+    @staticmethod
+    def _map_injection_anaesthesia(
+        list_item, list_fields
+    ) -> Optional[Anaesthetic]:
+        """Maps anaesthesic type, duration, level for Injection"""
+        anaesthetic_type = "isoflurane"
+        duration = list_item.get_property(
+            list_fields.FIRST_INJECTION_ISO_DURATION.value
+        )
+        level = list_item.get_property(list_fields.ROUND1_INJ_ISOLEVEL.value)
+        anaesthetic = Anaesthetic.construct(
+            type=anaesthetic_type,
+            duration=duration,
+            level=level,
+        )
+        return anaesthetic
+
+    def _map_list_item_to_injection(self, list_item: ClientObject):
+        """Maps a SharePoint ClientObject to an Injection model"""
+        list_fields = NeurosurgeryAndBehaviorList2019.ListField
+        start_date = parser.isoparse(
+            list_item.get_property(list_fields.DATE1ST_INJECTION.value)
+        ).date()
+        end_date = start_date
+        experimenter_full_name = list_item.get_property(
+            list_fields.LAB_TRACKS_REQUESTOR.value
+        )
+        iacuc_protocol = list_item.get_property(
+            list_fields.IACUC_PROTOCOL.value
+        )
+        animal_weight_prior = list_item.get_property(
+            list_fields.WEIGHT_BEFORE_SURGER.value
+        )
+        animal_weight_post = list_item.get_property(
+            list_fields.WEIGHT_AFTER_SURGERY.value
+        )
+        # TODO: all fields after this have diff field names for 2nd inj
+        anaesthesia = self._map_injection_anaesthesia(list_item, list_fields)
+        injection_duration = convert_str_to_time(
+            list_item.get_property(list_fields.INJ1_LENGHTOF_TIME.value)
+        )
+        recovery_time = list_item.get_property(
+            list_fields.FIRST_INJ_RECOVERY.value
+        )
+        workstation_id = list_item.get_property(
+            list_fields.WORK_STATION1ST_INJECTION.value
+        )
+        injection_type = list_item.get_property(list_fields.INJ1_TYPE.value)
+        injection_hemisphere = map_hemisphere(
+            list_item.get_property(list_fields.VIRUS_HEMISPHERE.value)
+        )
+        injection_coordinate_ml = parse_str_into_float(
+            list_item.get_property(list_fields.VIRUS_M_L.value)
+        )
+        # TODO: handle direction for coordinate ap
+        injection_coordinate_ap = parse_str_into_float(
+            list_item.get_property(list_fields.VIRUS_A_P.value)
+        )
+        # TODO: handle 2 values for depth (for now using 1st value)
+        injection_coordinate_depth = parse_str_into_float(
+            list_item.get_property(list_fields.VIRUS_D_V.value)
+        )
+        injection_angle = parse_str_into_float(
+            list_item.get_property(list_fields.INJ1ANGLE0.value)
+        )
+        if (
+                injection_type
+                == NeurosurgeryAndBehaviorList2019.InjectionType.IONTO.value
+        ):
+            instrument_id = list_item.get_property(
+                list_fields.IONTO_NUMBER_INJ1.value
+            )
+            injection_current = parse_str_into_float(
+                list_item.get_property(list_fields.INJ1_CURRENT.value)
+            )
+            alternating_current = list_item.get_property(
+                list_fields.INJ1_ALTERNATING_TIME.value
+            )
+            injection = IontophoresisInjection.construct(
+                start_date=start_date,
+                end_date=end_date,
+                experimenter_full_name=experimenter_full_name,
+                iacuc_protocol=iacuc_protocol,
+                animal_weight_prior=animal_weight_prior,
+                animal_weight_post=animal_weight_post,
+                anaesthesia=anaesthesia,
+                injection_duration=injection_duration,
+                recovery_time=recovery_time,
+                workstation_id=workstation_id,
+                instrument_id=instrument_id,
+                injection_hemisphere=injection_hemisphere,
+                injection_coordinate_ml=injection_coordinate_ml,
+                injection_coordinate_ap=injection_coordinate_ap,
+                injection_coordinate_depth=injection_coordinate_depth,
+                injection_angle=injection_angle,
+                injection_type=injection_type,
+                injection_current=injection_current,
+                alternating_current=alternating_current,
+            )
+        else:
+            instrument_id = list_item.get_property(
+                list_fields.NANOJECT_NUMBER_INJ10.value
+            )
+            injection_volume = parse_str_into_float(
+                list_item.get_property(list_fields.INJ1_VOL.value)
+            )
+            injection = NanojectInjection.construct(
+                start_date=start_date,
+                end_date=end_date,
+                experimenter_full_name=experimenter_full_name,
+                iacuc_protocol=iacuc_protocol,
+                animal_weight_prior=animal_weight_prior,
+                animal_weight_post=animal_weight_post,
+                anaesthesia=anaesthesia,
+                injection_duration=injection_duration,
+                recovery_time=recovery_time,
+                workstation_id=workstation_id,
+                instrument_id=instrument_id,
+                injection_hemisphere=injection_hemisphere,
+                injection_coordinate_ml=injection_coordinate_ml,
+                injection_coordinate_ap=injection_coordinate_ap,
+                injection_coordinate_depth=injection_coordinate_depth,
+                injection_angle=injection_angle,
+                injection_type=injection_type,
+                injection_volume=injection_volume,
+            )
+        return injection
+
+    @staticmethod
+    def _map_list_item_to_ophys_probe(list_item: ClientObject, list_fields):
+        """Maps a Sharepoint ListItem to a OphysProbe model"""
+        ophys_probes = []
+        # TODO: missing fields manufacturer, part_number, core_diameter
+        #  numerical_aperature, ferrule_material
+        fiber_implant1 = list_item.get_property(
+            list_fields.FIBER_IMPLANT1.value
+        )
+        if fiber_implant1:
+            name = ProbeName.PROBE_A.value
+            stereotactic_coordinate_ml = parse_str_into_float(
+                list_item.get_property(list_fields.VIRUS_M_L.value)
+            )
+            stereotactic_coordinate_ap = parse_str_into_float(
+                list_item.get_property(list_fields.VIRUS_A_P.value)
+            )
+            stereotactic_coordinate_dv = parse_str_into_float(
+                list_item.get_property(list_fields.FIBER_IMPLANT1_DV.value)
+            )
+            angle = parse_str_into_float(
+                list_item.get_property(list_fields.INJ1_ANGLE_V2.value)
+            )
+            ophys_probe1 = OphysProbe.construct(
+                name=name,
+                stereotactic_coordinate_ml=stereotactic_coordinate_ml,
+                stereotactic_coordinate_ap=stereotactic_coordinate_ap,
+                stereotactic_coordinate_dv=stereotactic_coordinate_dv,
+                angle=angle,
+            )
+            ophys_probes.append(ophys_probe1)
+        fiber_implant2 = list_item.get_property(
+            list_fields.FIBER_IMPLANT2.value
+        )
+        if fiber_implant2:
+            name = ProbeName.PROBE_B.value
+            stereotactic_coordinate_ml = parse_str_into_float(
+                list_item.get_property(list_fields.ML2ND_INJ.value)
+            )
+            stereotactic_coordinate_ap = parse_str_into_float(
+                list_item.get_property(list_fields.AP2ND_INJ.value)
+            )
+            stereotactic_coordinate_dv = parse_str_into_float(
+                list_item.get_property(list_fields.FIBER_IMPLANT2_DV.value)
+            )
+            angle = parse_str_into_float(
+                list_item.get_property(list_fields.INJ2_ANGLE_V2.value)
+            )
+            ophys_probe1 = OphysProbe.construct(
+                name=name,
+                stereotactic_coordinate_ml=stereotactic_coordinate_ml,
+                stereotactic_coordinate_ap=stereotactic_coordinate_ap,
+                stereotactic_coordinate_dv=stereotactic_coordinate_dv,
+                angle=angle,
+            )
+            ophys_probes.append(ophys_probe1)
+        return ophys_probes
+
+    def _map_list_item_to_fiber_implant(
+            self,
+            list_item: ClientObject,
+    ) -> FiberImplant:
+        """Maps a SharePoint ListItem to a FiberImplant model"""
+        list_fields = NeurosurgeryAndBehaviorList2019.ListField
+        start_date = parser.isoparse(
+            list_item.get_property(list_fields.DATE_OF_SURGERY.value)
+        ).date()
+        end_date = start_date
+        experimenter_full_name = list_item.get_property(
+            list_fields.LAB_TRACKS_REQUESTOR.value
+        )
+        iacuc_protocol = list_item.get_property(
+            list_fields.IACUC_PROTOCOL.value
+        )
+        animal_weight_prior = list_item.get_property(
+            list_fields.WEIGHT_BEFORE_SURGER.value
+        )
+        animal_weight_post = list_item.get_property(
+            list_fields.WEIGHT_AFTER_SURGERY.value
+        )
+        probes = self._map_list_item_to_ophys_probe(list_item, list_fields)
+        fiber_implant = FiberImplant.construct(
+            start_date=start_date,
+            end_date=end_date,
+            experimenter_full_name=experimenter_full_name,
+            iacuc_protocol=iacuc_protocol,
+            animal_weight_prior=animal_weight_prior,
+            animal_weight_post=animal_weight_post,
+            probes=probes,
+        )
+        return fiber_implant
+
+    @staticmethod
+    def _map_hp_anaesthesia(list_item, list_fields) -> Optional[Anaesthetic]:
+        """Maps anaesthesic type, duration, level for HP/craniotomy"""
+        anaesthetic_type = "isoflurane"
+        # TODO: duration
+        level = list_item.get_property(list_fields.HP_ISO_LEVEL.value)
+        anaesthetic = Anaesthetic.construct(
+            type=anaesthetic_type,
+            level=level,
+        )
+        return anaesthetic
+
+    @staticmethod
+    def _map_craniotomy_type(sp_craniotomy_type) -> Optional[CraniotomyType]:
+        """Maps craniotomy type"""
+        CT = NeurosurgeryAndBehaviorList2019.CraniotomyType
+        if sp_craniotomy_type:
+            if sp_craniotomy_type == CT.VISUAL_CORTEX.value:
+                return CraniotomyType.VISCTX
+            elif sp_craniotomy_type == CT.FRONTAL_WINDOW.value:
+                return CraniotomyType.THREE_MM
+            elif sp_craniotomy_type in {
+                CT.WHC_NP.value,
+                CT.WHC_2P.value,
+            }:
+                return CraniotomyType.WHC
+            else:
+                return CraniotomyType.OTHER
+        return None
+
+    def _map_list_item_to_craniotomy(
+            self, list_item: ClientObject
+    ) -> Craniotomy:
+        """Maps a SharePoint ListItem to a Craniotomy model"""
+        # TODO: missing fields (implant_part_number, protective_material)
+        list_fields = NeurosurgeryAndBehaviorList2019.ListField
+        start_date = parser.isoparse(
+            list_item.get_property(list_fields.DATE_OF_SURGERY.value)
+        ).date()
+        end_date = start_date
+        experimenter_full_name = list_item.get_property(
+            list_fields.LAB_TRACKS_REQUESTOR.value
+        )
+        iacuc_protocol = list_item.get_property(
+            list_fields.IACUC_PROTOCOL.value
+        )
+        animal_weight_prior = list_item.get_property(
+            list_fields.WEIGHT_BEFORE_SURGER.value
+        )
+        animal_weight_post = list_item.get_property(
+            list_fields.WEIGHT_AFTER_SURGERY.value
+        )
+        anaesthesia = self._map_hp_anaesthesia(list_item, list_fields)
+        craniotomy_type = self._map_craniotomy_type(
+            list_item.get_property(list_fields.CRANIOTOMY_TYPE.value)
+        )
+        # TODO: handle size and coords by craniotomy_type ?
+        craniotomy_hemisphere = map_choice(
+            list_item.get_property(list_fields.HP_LOC.value)
+        )
+        craniotomy_coordinates_ml = parse_str_into_float(
+            list_item.get_property(list_fields.HP_M_L.value)
+        )
+        craniotomy_coordinates_ap = parse_str_into_float(
+            list_item.get_property(list_fields.HP_A_P.value)
+        )
+        craniotomy_size = parse_str_into_float(
+            list_item.get_property(list_fields.HP_DIAMETER.value)
+        )
+        dura_removed = convert_str_to_bool(
+            list_item.get_property(list_fields.HP_DUROTOMY.value)
+        )
+        workstation_id = list_item.get_property(
+            list_fields.HP_WORK_STATION.value
+        )
+        craniotomy = Craniotomy.construct(
+            start_date=start_date,
+            end_date=end_date,
+            experimenter_full_name=experimenter_full_name,
+            iacuc_protocol=iacuc_protocol,
+            animal_weight_prior=animal_weight_prior,
+            animal_weight_post=animal_weight_post,
+            anaesthesia=anaesthesia,
+            craniotomy_type=craniotomy_type,
+            craniotomy_hemisphere=craniotomy_hemisphere,
+            craniotomy_coordinates_ml=craniotomy_coordinates_ml,
+            craniotomy_coordinates_ap=craniotomy_coordinates_ap,
+            craniotomy_size=craniotomy_size,
+            dura_removed=dura_removed,
+            workstation_id=workstation_id,
+        )
+        return craniotomy
+
+    @staticmethod
+    def _map_headpost_type(headpost_type: str):
+        """Maps Sharepoint HeadPostType to fields in HeadFrame model"""
+        if (
+                headpost_type
+                == NeurosurgeryAndBehaviorList2019.HeadPostType.CAM.value
+        ):
+            headframe_type = "CAM-style"
+            headframe_part_number = "0160-100-10 Rev A"
+            well_type = "CAM-style"
+            well_part_number = None
+        elif (
+                headpost_type
+                == NeurosurgeryAndBehaviorList2019.HeadPostType.NEUROPIXEL.value
+        ):
+            headframe_type = "Neuropixel-style"
+            headframe_part_number = "0160-100-10"
+            well_type = "Neuropixel-style"
+            well_part_number = "0160-200-36"
+        elif (
+                headpost_type
+                == NeurosurgeryAndBehaviorList2019.HeadPostType.MESO_NGC.value
+        ):
+            headframe_type = "NGC-style"
+            headframe_part_number = "0160-100-10"
+            well_type = "Mesoscope-style"
+            well_part_number = "0160-200-20"
+        elif (
+                headpost_type
+                == NeurosurgeryAndBehaviorList2019.HeadPostType.WHC_NP.value
+        ):
+            headframe_type = "WHC #42"
+            headframe_part_number = "42"
+            well_type = "Neuropixel-style"
+            well_part_number = "0160-200-36"
+        elif (
+                headpost_type
+                == NeurosurgeryAndBehaviorList2019.HeadPostType.NGC.value
+        ):
+            headframe_type = "NGC-style"
+            headframe_part_number = "0160-100-10"
+            well_type = None
+            well_part_number = None
+        elif (
+                headpost_type
+                == NeurosurgeryAndBehaviorList2019.HeadPostType.AI_HEADBAR.value
+        ):
+            headframe_type = "AI Straight Headbar"
+            headframe_part_number = None
+            well_type = None
+            well_part_number = None
+        else:
+            return None, None, None, None
+        return (
+            headframe_type,
+            headframe_part_number,
+            well_type,
+            well_part_number,
+        )
+
+    def _map_list_item_to_head_frame(
+            self, list_item: ClientObject
+    ) -> Headframe:
+        """Maps a SharePoint ListItem to a HeadFrame model"""
+        list_fields = NeurosurgeryAndBehaviorList2019.ListField
+        start_date = parser.isoparse(
+            list_item.get_property(list_fields.DATE_OF_SURGERY.value)
+        ).date()
+        end_date = start_date
+        experimenter_full_name = list_item.get_property(
+            list_fields.LAB_TRACKS_REQUESTOR.value
+        )
+        iacuc_protocol = list_item.get_property(
+            list_fields.IACUC_PROTOCOL.value
+        )
+        animal_weight_prior = parse_str_into_float(
+            list_item.get_property(list_fields.WEIGHT_BEFORE_SURGER.value)
+        )
+        animal_weight_post = parse_str_into_float(
+            list_item.get_property(list_fields.WEIGHT_AFTER_SURGERY.value)
+        )
+        anaesthesia = self._map_hp_anaesthesia(list_item, list_fields)
+        headpost_type = list_item.get_property(list_fields.HEADPOST_TYPE.value)
+        (
+            headframe_type,
+            headframe_part_number,
+            well_type,
+            well_part_number,
+        ) = self._map_headpost_type(headpost_type)
+        head_frame = Headframe.construct(
+            start_date=start_date,
+            end_date=end_date,
+            experimenter_full_name=experimenter_full_name,
+            iacuc_protocol=iacuc_protocol,
+            animal_weight_prior=animal_weight_prior,
+            animal_weight_post=animal_weight_post,
+            anaesthesia=anaesthesia,
+            headframe_type=headframe_type,
+            headframe_part_number=headframe_part_number,
+            well_type=well_type,
+            well_part_number=well_part_number,
+        )
+        return head_frame
